requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Data validation
pydantic==2.7.1
pydantic-settings==2.3.1

# Environment variables
python-dotenv==1.0.1

# LangChain
langchain==0.2.6
langchain-openai==0.1.14
langchain-core==0.2.10

# HTTP client (used by langchain-openai)
httpx==0.27.0
